solution

+++ src/main/kotlin/BoxVolumeCalculator.kt
@@ -1,3 +1,3 @@
-fun getBoxVolume(length: Double, width: Double, height: Double = 1.0): Double {
+fun getBoxVolume(length: Int, width: Int, height: Int = 1): Int {
     return length * width * height
 }
